🧪 Refactor and test parseFrontmatter edge cases
